feat(aluguel) implementado nova funcionalidade de alguar jogo e ver historico

+++ 2948-kotlin-alugames-curso1-aula02/src/main/kotlin/br/com/alura/davgame/modelo/InfoJogo.kt
@@ -4,7 +4,7 @@ import com.google.common.base.MoreObjects
 
 data class InfoJogo(val titulo: String,
                     val capa: String,
-                    val preco: Float,
+                    val preco: Double,
                     val descricao: String) {
 
 }
